+++ src/main/java/com/bootcamp/cdd/request/CupomRequest.java
@@ -15,11 +15,11 @@ public class CupomRequest {
     @UniqueValue(fieldName = "codigo", domainClass = Cupom.class)
     private String codigo;
     @NotNull(message = "A porcentagem precisa ser preenchida") @Positive(message = "O valor precisa ser positivo")
-    private BigDecimal porcentagem;
+    private double porcentagem;
     @Future(message = "a data precisa ser no futuro")
     private LocalDate dataValidade;
 
-    public CupomRequest(@NotBlank(message = "O codigo precisa ser preenchido") String codigo, @NotNull(message = "A porcentagem precisa ser preenchida") @Positive(message = "O valor precisa ser positivo") BigDecimal porcentagem, @Future(message = "a data precisa ser no futuro") LocalDate dataValidade) {
+    public CupomRequest(@NotBlank(message = "O codigo precisa ser preenchido") String codigo, @NotNull(message = "A porcentagem precisa ser preenchida") @Positive(message = "O valor precisa ser positivo") double porcentagem, @Future(message = "a data precisa ser no futuro") LocalDate dataValidade) {
         this.codigo = codigo;
         this.porcentagem = porcentagem;
         this.dataValidade = dataValidade;
